Try to fix Arbitrage strategy's default amount bug
(WIP)

There should not only be 1 default value for the
arbitrage strategy's amount, but several:

Consider a hypothetical situation with a DOT/USDT
trading pair.

If the price of the trading pair is too far from 1 and
the default amount per trade is set at either
0.1 (to account for USDT) or 0.01 (to account for DOT),
e.g. 10 DOT/USDT, then in one direction, the
trades will be too small; in the other direction, they
will be too large..