Perform two price requests simultaneously (WIP)
On careful consideration, this will not work.
Submitting both price estimation Move calls in the same PTB
means that if one of them fails, the other won't
get to return a result.

This has the unfortunate consequence that RAMM price estimations
may require 2 RPC calls, when the assets involved have imbalance
ratios close to the RAMM's threshold (ONE +/- DELTA).